Fix bug tolerance sospsimplfiy.
There was an issue with sospsimplify, wherein rows of A were extracted which contain exactly 2 nonzero elements, not relying on any tolerance, but then the associated nonzero columns were extracted with a tolerance, producing less column indices than expected. This has been fixed by removing the tolerance in the extraction of column indices as well.