Fixed missing 'tol' issue in sossolve.
When 'sossolve' was called with the 'psimplify' option, it did not explicitly check if a tolerance was specified, only if a field 'options.params' are specified. This produced an error in case a field 'options.params' is specified but no 'options.params.tol' was specified. Currently solved by adding default tolerance in case 'psimplify' is used. It's possible other parts of the code may also fail if no 'options.params.tol' field is specified, but so far I haven't encountered any errors.
Also, added the option to specify 'options.simiplfy=0' to turn off 'psimplify'.