Fixed issue with 'pvar' option in sospolymatrixvar.
The 'pvar' option called 'sospolyvar' with 3 arguments, the third of which specifying the 'pvar' option. However, 'sospolyvar' takes this option as its fourth argument. I corrected to pass an empty field as third argument and 'pvar' as fourth, solving the problem. However, maybe 'sospolyvar' needs to be adjusted instead, to allow 'pvar' to be specified as third argument as well, since only the first two arguments of 'sospolyvar' are actually mandatory.