Fixed issues specifying dpvar objective.
The function did not allow dpvars to be specified unless their 'varname' field was of size [0,0], so that no independent variables would be included in the objective. However, the second dimension of the 'varname' field should always be 1, so I adjusted to verify only the first dimension to be zero.
In addition, the 'sos.objective' field produced in the 'dpvar' case was a row vector, producing errors later on (in 'sossolve'). This has been fixed to now produces a column vector.