Include interval case for dpvar.
The function allows an interval to be specified when 'syms' implementation is used. However, in the 'dpvar' case, this interval is ignored, without warning the user. The function has been adjusted to allow [-Inf,b] and [a,Inf] intervals to be specified, and return an error if a finite interval [a,b] is specified, in the 'dpvar' case.