Adjust dpvar getequation for Mineq case.
For a matrix-valued dpvar class expression, "getequation" produces a separate set of monomials for each element of the matrix, discarding monomials that do not contribute. However, for constraints of type 'ineq', the resulting "prog.expr" format returned by "getequation" does not allow for proper processing/conversion of the constraint in "sossolve". The function has been adjusted to take an additional optional input 'ineq', so that, when this option is passed, "getequation" returns a single set of unique monomials to be used for all elements of the matrix-valued object.