Adjust dpvar sosconstr for Mineq case.
For constraints of type 'ineq', passing a matrix-valued dpvar class object to "getequation" resulted in an expression "prog.expr" that could not be properly processed. The function "getequation" has been adjusted to take an optional input 'ineq', producing an output that can be processed correctly. The function "sosconstr" has been adjusted accordingly, passing the constraint type to "getequation" when processing dpvar class objects.