Adjusted dpvar "addextrasosvar" in sossolve.
Previously, imposing inequality constraints on matrix-valued dpvar objects using the 'Mineq' option produced an expression in the program structure that could not be properly processed. Therefore, "getequation" and "sosconstr" have been adjusted to return a different format when 'ineq' constraints are imposed. An alternative version of the "addextrasosvar" subroutine in "sossolve" has now been implemented for the dpvar case, to process the 'ineq' constraints in the new format.
Note: The alternative subroutine is also called in the pvar case, despite the fact that the pvar expression format is not suited for this new way of processing. We may need to make some adjustments to be able to distinguish pvar and dpvar implementations from the program structure.