Add feasibility check to syms addextrasosvar.
In the "addextrasosvar" subroutine of "sossolve", an SOS variable S(x) is constructed to implement the constraint F(x)=S(x) for a given polynomial F(x). In doing so, a minimal basis Z(x) of monomials is establish such that S(x) is SOS in Z(X) and can match all terms in F(x). However, if F(x) contains only odd degree monomials, this minimal basis will be empty, producing an error. To avoid this, a check has been implemented to stop the solving process if the monomial vector Z(x) is found to be empty, warning the user that the problem is infeasible in this case.