Convert "double" to appropriate data class.
When calling "sosconstr" with an expression that is just a constant matrix of class "double", the function previously returned an error. The function has been adjusted to convert objects of type "double" to "sym" class objects if the sym implementation is used, and "dpvar" class objects otherwise. Furthermore, expressions of type "polynomial" are now also converted to class "dpvar" objects before calling "getequation". This is because the updated "addextrasosvar" subroutine in "sossolve" cannot distinguish between constraints imposed using "dpvar" and "polynomial"  expressions, and so both expressions must be implemented in the same format. However, in the reversionary implementation where decision variables are implemented as 'pvar' objects, the implementation is no longer truly reversionary, and computation time may be increased.