Fix bug syms "getequation".
In lines 309 and 325 of the code, an element-wise transpose of a sym expression is computed. However, this transpose does not appear to produce the expected result, likely due to an update in the symbolic toolbox. The transpose has now been excluded, avoiding the error, although other problems may still be present.