Fix missing 'feasextrasos' sossolve.
In the last update to "sossolve", a check was added, setting the value of an object "feasextrasos" to 0 if a problem is found to have obviously infeasible inequality constraints. However, the value of "feasextrasos" was not specified if no inequality constraints appeared in the problem, producing an error. This has been resolved, setting a default value "feasextrasos=1" before processing the inequality constraints.