Fix issue empty "K.s" sossolve.
When calling "sossolve" with solvers 'sdpt3', 'sdpnal' and 'sdpnalplus', the function checked whether any SOS variables are present in the problem by verifying "K.s(1)~=0", where the field "K.s" contains the size of each SOS variable in the problem. However, a lack of SOS variables results in an empty field "K.s", rather than the first element being equal to zero, producing an error. A check has been added to ensure the field is not empty, solving the issue. However, the "K.s(1)~=0" is still included as well, as an SOS variable of size 0 would certainly produce an error as well, though I don't think such a variable can actually be implemented.