Allow matrix-valued linear indices in subsref.
The function previously allowed linear indices to a dpvar object to be specified as an mx1 or 1xm array, always returning an mx1 array as output. The function has been updated to allow linear indices to be specified as any mxn array, producing a dpvar output with the same dimensions as the indices.