Fix issues edge cases findbound.
When the function was called with a 'polynomial' class object for the equality or inequality constraints, an error occurred if the "degmat" field of this polynomial was empty. In particular, the function was unable to establish a maximal degree of the polynomial in this case, resulting in issues. This problem has been resolved, setting the maximal degree to zero in this case.
In addition, a 'dpvar' implementation of the code has been added, allowing the function for which to find a bound to be specified as a 'dpvar' class object. Note, however, that in general it does not make sense for the function to bound to have decision variables itself.