Fix output dimensions empty concatenation.
In "horzcat.m" and "vertcat.m", when both inputs to concatenate are empty polynomials, the function would previously just return one of the polynomials. However, empty objects may still have a nonzero row dimension, or a nonzero column dimension, in which case the row or column dimension of the output would have been incorrect. For example, vertically concatenating a 1x0 object and a 2x0 object, a 3x0 object would be expected, but the "vertcat" function would previously return either the 1x0 or the 2x0 object. This issue has been resolved, ensuring that the row dimension of the output of "vertcat" matches the sum of the row dimensions of the input, and similarly for the column dimensions in "horzcat".
NOTE: If only one of the input polynomials is empty, the function still does not check that the column dimensions (in "vertcat") or row dimensions (in "horzcat") of the two inputs match. That is, [(zeros(0,10); [1,x]] for some 'pvar' object x would simply return [1,x]. I'm not quite sure if this feature is useful, or if it may lead to issues, so I have left it as is for now.