Fix rational pvar/dpvar implemenation of findsos.
The pvar/dpvar implementation of the "findsos" function with "flag='rational' was simply a copy of the syms implementation, producing an error as the function "expand" does not exist for polynomial class objects. This has been resolved. In addition, some more checks have been added to avoid rough exists/errors when the input function has non-integer coefficients.