Bugfix sossolve sdpa case.
The sedumiwrap for solver 'sdpa' does not currently accept fields "K.q" and "K.r". If these are zero, however, the function should be able to run. A check has been added to determine whether these values are zero, and remove them from "K", so that the sedumiwrap can be properly run.