Minor bugfix findsos.
If the function was called with more than 2 outputs, but the function terminated prematurely due to e.g. an SOS decomposition not being possible, an error would be thrown as the last two function outputs were not specified. This has been resolved by also initializing the additional outputs as empty fields when the function terminates prematurely.