Bugfix "sossolve" using Mosek.
Using the 'psimplify' option (`psimplify=1`), in some cases, certain decision variables are found not to contribute at all to the optimization program before calling the solver. This is indicated by setting K.s(i)=0 for this ith matrix-valued variable, so that the ith matrix-valued variable is interpreted to be of dimension 0x0. However, Mosek does not accept 0x0 dimensional variables, producing an error. This issue has been resolved, setting `K.s(K.s==0)=[]`, thus removing the 0x0-dimensional variables from the SDP structure altogether. In the conversion back to SOSP solutions, this is interpreted the same way as keeping K.s(i)=0.