Bugfix findsos.
When the proposed polynomial `P` contained elements on the diagonal which are constant, the function ran into an issue with an empty field `Pjj.degmat`. This issue has been resolved. In addition, a feature has been added to allow constant matrices (of type 'double') to be passed as well, in so far as that is useful.