Update sospolyvar_mat.
The function requires row and column dimensions to be specified as a 1x2 array in the third argument. However, in many cases, the row and column dimensions will be equal, and users may only provide one dimension. Accordingly, the function has been updated to allow a single dimension to be specified as the third argument, setting both the row and column dimension equal to this value.