Update polynomial/power.
Functionality update: the function now allows matrix-valued ```polynomial``` objects to be raised to vector-valued powers (i.e. Matrix.^Vector) and vice versa (Vector.^Matrix), so long as either the number of rows or the number of columns of the matrix and the vector match. The vector is concatenated with itself (using ```repmat```) to match the dimensions of the matrix, at which point element-wise powers are computed in the standard manner.