Update sospsimplify.
A tolerance has been re-introduced in simplifying the SOS program using ```sospsimplify```. Coefficients below the tolerance are assumed to be zero in eliminating constraints. Without this tolerance, there are cases where a feasible SOS program may be reduced to an infeasible one. The value of the tolerance can be set using ```opts.simplify_tol=...`` in calling ```sossolve(prog,opts)```, and defaults to 1e-12 if not specified.