Create 'numel' functions.
It appears no 'numel' functions had yet been created for 'polynomial' and 'dpvar' class objects, so that calling `numel(P)` for such an object returned an erroneous value. Although this does not seem to have caused any major errors, there was a minor issue in `getequation` where `numel` was used to determine the number of independent variables in the optimization program. As a result, the number of columns of the output monomial degree matrix `Z` did not match the true number of variables in the optimization program, making it unclear what variables these degrees belonged to. This has been resolved.