Update sospsimplify.m
There was an issue in the case the `cxUz`, `cxUp` or `cxUn` was empty, wherein it was assigned a 0x0 rather than 0x2 array. This caused problems in evaluating e.g. `cxUz(:,1)` later on. This has been resolved, concatenating each array with a 0x2 array to ensure the array has two columns (even if it has zero rows).