Update sospolyvar_mat.m
There was an issue in the function where the variable names of the monomial vector `ZSym` were compared with the actual ('polynomial'-class) variables in the program structure `sos`, rather than the names of those variables. This has been resolved. In addition, a check has been added to return an unmodified program structure if the dimensions of the desired matrix variable are zero, to avoid adding empty variables to the program structure. Finally, a check has been added to throw an error if the user tries calling the function for a program declared using the 'syms' format.