Update sossolve.m
If no SDP solver is specified, the function now uses Mosek as first choice of default solver, if encountered on the Matlab path. Since Mosek requires a license to use, it is likely that if users have Mosek on the path, then that is also their preferred solver of choice.  If Mosek is not on the path, the function checks for other SDP solvers, or throws an error if no SDP solver is encountered.