Update diff.m
Add support for higher order derivatives, e.g. `diff(f,x,2)` to evaluate the second order derivative of `f` with respect to `x`. Implementation is based on a simple for-loop, repeatedly applying the `jacobian` function, which may not be the most efficient approach. However, it saves the hassle of evaluating this derivative manually as `diff(diff(f,x),x))`.